requirements.txt
python-telegram-bot==20.3
vobject==0.9.6.1